Added PeopleFund.it page
A big thanks to all of the people who donated to YRS2012 through
peoplefund.it -- and were in the categories where they 'qualified' for
a thanks on the site.

Hopefully excludes those who opted out -- although that's hard to grok
from the PF website. Or maybe only a few people asked to be anon...?